Run the installer in a sub-sandbox
Since 2c9d2084, the desktop folder is exposed to the app. Due to this, if the
corresponding option in the installer is selected, the installer creates a set
of desktop files in the user's desktop.

These files are not needed (as the flatpak app aready exposes the necessary
desktop files) and do not work, as they refer to the original binaries names.

With this commit the installer is launched in a restricted sub-sandbox, thanks
to `flatpak-spawn --sandbox ...`, exposing only the strictly-needed directories.